fix 修改 linuxdo 网站用户注册时赠送的金额

+++ model/linuxdo.go
@@ -15,7 +15,7 @@ func (user User) QuotaForNewLinuxDoUser() {
 	}else  {
 		linuxDoLevel = float64(user.LinuxDoLevel) - 1;
 	}
-	var quota = int(config.QuotaPerUnit) * int(linuxDoLevel)
-	_ = IncreaseUserQuota(user.Id, quota)
-	RecordLog(user.Id, LogTypeSystem, fmt.Sprintf("linux do %v级用户注册, 赠送 %s", user.LinuxDoLevel, common.LogQuota(quota)))
+	var quota = float64(config.QuotaPerUnit) * linuxDoLevel
+	_ = IncreaseUserQuota(user.Id, int(quota))
+	RecordLog(user.Id, LogTypeSystem, fmt.Sprintf("linux do %v级用户注册, 赠送 %s", user.LinuxDoLevel, common.LogQuota(int(quota))))
 }
